logic moved to firebase
